move get_SQLAlchemyDriver to config
